feat: LinkedList isEmpty test
