--- Calculator/CalculatorItemQueueLinkedListTests/CalculatorItemQueueLinkedListTests.swift
@@ -1,24 +1,0 @@
-//
-//  CalculatorItemQueueLinkedListTests.swift
-//  CalculatorItemQueueLinkedListTests
-//
-//  Created by 김진주 on 2023/10/03.
-//
-
-import XCTest
-@testable import Calculator
-
-final class CalculatorItemQueueLinkedListTests: XCTestCase {
-    var sut: CalculatorItemQueueLinkedList<Any>!
-
-    override func setUpWithError() throws {
-        try super.setUpWithError()
-        sut = CalculatorItemQueueLinkedList<Any>()
-    }
-
-    override func tearDownWithError() throws {
-        try super.tearDownWithError()
-        sut = nil
-    }
-
-}
